+++ src/Registration/DecoratorRepository.cs
@@ -17,7 +17,8 @@ namespace Stashbox.Registration
         {
             RegistrationSelectionRules.GenericFilter,
             RegistrationSelectionRules.ConditionFilter,
-            RegistrationSelectionRules.ScopeNameFilter
+            RegistrationSelectionRules.ScopeNameFilter,
+            RegistrationSelectionRules.DecoratorFilter,
         };
 
         public void AddDecorator(Type type, ServiceRegistration serviceRegistration, bool remap)
@@ -38,6 +39,9 @@ namespace Stashbox.Registration
         public IEnumerable<ServiceRegistration> GetDecoratorsOrDefault(Type implementationTypeToDecorate, TypeInformation typeInformation, ResolutionContext resolutionContext) =>
             this.GetRegistrationsForType(typeInformation.Type)?.FilterInclusiveOrDefault(typeInformation.CloneForType(implementationTypeToDecorate), resolutionContext, this.filters);
 
+        public ServiceRegistration GetNextDecoratorOrDefault(Type implementationTypeToDecorate, TypeInformation typeInformation, ResolutionContext resolutionContext) =>
+            this.GetRegistrationsForType(typeInformation.Type)?.SelectOrDefault(typeInformation, resolutionContext, this.filters);
+
         public IEnumerable<KeyValuePair<Type, ServiceRegistration>> GetRegistrationMappings() =>
             repository.Walk().SelectMany(reg => reg.Value.Select(r => new KeyValuePair<Type, ServiceRegistration>(reg.Key, r)));
 
@@ -48,5 +52,6 @@ namespace Stashbox.Registration
                 ? registrations
                 : repository.GetOrDefault(type.GetGenericTypeDefinition());
         }
+
     }
 }