+++ src/Resolution/ResolutionStrategy.cs
@@ -47,9 +47,9 @@ namespace Stashbox.Resolution
                     }
                 }
 
-                var decorators = resolutionContext.Decorators.GetOrDefault(typeInformation.Type, true);
-                if (decorators != null)
-                    return this.BuildExpressionForDecorator(decorators.Pop(),
+                var decorators = resolutionContext.RemainingDecorators.GetOrDefault(typeInformation.Type, true);
+                if (decorators != null && decorators.Length > 0)
+                    return this.BuildExpressionForDecorator(decorators.Front(),
                         resolutionContext.BeginDecoratingContext(typeInformation.Type, decorators), typeInformation.Type, decorators);
             }
 
@@ -80,14 +80,14 @@ namespace Stashbox.Resolution
 
             return registrations.Select(reg =>
             {
-                var decorators = resolutionContext.Decorators.GetOrDefault(typeInformation.Type, true);
-                if (decorators == null)
+                var decorators = resolutionContext.RemainingDecorators.GetOrDefault(typeInformation.Type, true);
+                if (decorators == null || decorators.Length == 0)
                     return this.BuildExpressionForRegistration(reg, resolutionContext, typeInformation);
 
                 decorators.ReplaceBack(reg);
-                return this.BuildExpressionForDecorator(decorators.Pop(),
-                    resolutionContext.BeginDecoratingContext(typeInformation.Type, decorators), typeInformation.Type,
-                    decorators);
+                return this.BuildExpressionForDecorator(decorators.Front(),
+                    resolutionContext.BeginDecoratingContext(typeInformation.Type, decorators), 
+                    typeInformation.Type, decorators);
             });
         }
 
@@ -108,7 +108,7 @@ namespace Stashbox.Resolution
 
             var stack = decorators.AsStack();
             stack.PushBack(serviceRegistration);
-            return this.BuildExpressionForDecorator(stack.Pop(),
+            return this.BuildExpressionForDecorator(stack.Front(),
                 resolutionContext.BeginDecoratingContext(requestedType, stack), requestedType, stack);
         }