+++ test/ContainerTests.cs
@@ -6,7 +6,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using Stashbox;
 using Xunit;
 using System.Reflection;