+++ test/NamedResolveTests.cs
@@ -1,6 +1,5 @@
 ﻿using Stashbox.Exceptions;
 using System;
-using System.Linq;
 using System.Collections;
 using System.Reflection;
 using System.Text;